@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1000
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -26,38 +26,31 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   auto_size_text: 3.0.0
-  built_collection: 5.1.1
-  built_value: 8.4.3
   cached_network_image: 3.2.1
-  cloud_firestore: 4.2.0
-  cloud_firestore_platform_interface: 5.9.1
-  cloud_firestore_web: 3.1.1
-  equatable: 2.0.3
-  firebase_auth: 4.2.0
-  firebase_auth_platform_interface: 6.11.4
-  firebase_auth_web: 5.2.0
-  firebase_core: 2.4.0
-  firebase_core_platform_interface: 4.5.2
-  firebase_core_web: 2.0.2
+  cloud_firestore: 4.8.0
+  cloud_firestore_platform_interface: 5.15.0
+  cloud_firestore_web: 3.6.0
+  equatable: 2.0.5
+  firebase_auth: 4.6.2
+  firebase_auth_platform_interface: 6.15.2
+  firebase_auth_web: 5.5.2
+  firebase_core: 2.13.1
+  firebase_core_platform_interface: 4.8.0
+  firebase_core_web: 2.5.0
   flutter_animate: 4.1.1+1
   flutter_cache_manager: 3.3.0
+  flutter_html: 3.0.0-alpha.5
   font_awesome_flutter: 10.1.0
   from_css_color: 2.0.0
-  geolocator: 9.0.1
-  geolocator_android: 4.1.0
-  geolocator_apple: 2.2.1
-  geolocator_platform_interface: 4.0.6
-  geolocator_web: 2.1.6
-  go_router: 3.1.0
+  go_router: 7.1.1
   google_fonts: 4.0.3
   google_sign_in: 6.0.2
   google_sign_in_android: 6.1.8
   google_sign_in_ios: 5.6.1
   google_sign_in_platform_interface: 2.4.0
   google_sign_in_web: 0.11.0+2
-  intl: 0.17.0
+  intl: 0.18.0
   json_path: 0.4.1
-  json_serializable: 6.6.1
   mime_type: 1.0.0
   page_transition: 2.0.4
   path_provider: 2.0.14
@@ -76,6 +69,7 @@ dependencies:
   sign_in_with_apple_platform_interface: 1.0.0
   sign_in_with_apple_web: 1.0.1
   sqflite: 2.2.6
+  stream_transform: 2.1.0
   timeago: 3.2.2
   url_launcher: 6.1.10
   url_launcher_android: 6.0.27
@@ -88,9 +82,6 @@ dependencies:
   cupertino_icons: ^1.0.0
 
 dev_dependencies:
-  analyzer: 5.10.0
-  build_runner: 2.3.3
-  built_value_generator: 8.4.3
 
   flutter_test:
     sdk: flutter
