@@ -7,7 +7,7 @@ import '../auth_manager.dart';
 import '../base_auth_user_provider.dart';
 import '../../flutter_flow/flutter_flow_util.dart';
 
-import '../../backend/backend.dart';
+import '/backend/backend.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:stream_transform/stream_transform.dart';
 import 'anonymous_auth.dart';
@@ -16,8 +16,32 @@ import 'email_auth.dart';
 import 'firebase_user_provider.dart';
 import 'google_auth.dart';
 import 'jwt_token_auth.dart';
+import 'github_auth.dart';
 
 export '../base_auth_user_provider.dart';
+
+class FirebasePhoneAuthManager extends ChangeNotifier {
+  bool? _triggerOnCodeSent;
+  FirebaseAuthException? phoneAuthError;
+  // Set when using phone verification (after phone number is provided).
+  String? phoneAuthVerificationCode;
+  // Set when using phone sign in in web mode (ignored otherwise).
+  ConfirmationResult? webPhoneAuthConfirmationResult;
+  // Used for handling verification codes for phone sign in.
+  void Function(BuildContext)? _onCodeSent;
+
+  bool get triggerOnCodeSent => _triggerOnCodeSent ?? false;
+  set triggerOnCodeSent(bool val) => _triggerOnCodeSent = val;
+
+  void Function(BuildContext) get onCodeSent =>
+      _onCodeSent == null ? (_) {} : _onCodeSent!;
+  set onCodeSent(void Function(BuildContext) func) => _onCodeSent = func;
+
+  void update(VoidCallback callback) {
+    callback();
+    notifyListeners();
+  }
+}
 
 class FirebaseAuthManager extends AuthManager
     with
@@ -31,6 +55,7 @@ class FirebaseAuthManager extends AuthManager
   String? _phoneAuthVerificationCode;
   // Set when using phone sign in in web mode (ignored otherwise).
   ConfirmationResult? _webPhoneAuthConfirmationResult;
+  FirebasePhoneAuthManager phoneAuthManager = FirebasePhoneAuthManager();
 
   @override
   Future signOut() {
@@ -115,20 +140,47 @@ class FirebaseAuthManager extends AuthManager
       _signInOrCreateAccount(context, googleSignInFunc, 'GOOGLE');
 
   @override
+  Future<BaseAuthUser?> signInWithGithub(BuildContext context) =>
+      _signInOrCreateAccount(context, githubSignInFunc, 'GITHUB');
+
+  @override
   Future<BaseAuthUser?> signInWithJwtToken(
-          BuildContext context, String jwtToken) =>
+    BuildContext context,
+    String jwtToken,
+  ) =>
       _signInOrCreateAccount(context, () => jwtTokenSignIn(jwtToken), 'JWT');
+
+  void handlePhoneAuthStateChanges(BuildContext context) {
+    phoneAuthManager.addListener(() {
+      if (!context.mounted) {
+        return;
+      }
+
+      if (phoneAuthManager.triggerOnCodeSent) {
+        phoneAuthManager.onCodeSent(context);
+        phoneAuthManager
+            .update(() => phoneAuthManager.triggerOnCodeSent = false);
+      } else if (phoneAuthManager.phoneAuthError != null) {
+        final e = phoneAuthManager.phoneAuthError!;
+        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+          content: Text('Error: ${e.message!}'),
+        ));
+        phoneAuthManager.update(() => phoneAuthManager.phoneAuthError = null);
+      }
+    });
+  }
 
   @override
   Future beginPhoneAuth({
     required BuildContext context,
     required String phoneNumber,
-    required VoidCallback onCodeSent,
+    required void Function(BuildContext) onCodeSent,
   }) async {
+    phoneAuthManager.update(() => phoneAuthManager.onCodeSent = onCodeSent);
     if (kIsWeb) {
-      _webPhoneAuthConfirmationResult =
+      phoneAuthManager.webPhoneAuthConfirmationResult =
           await FirebaseAuth.instance.signInWithPhoneNumber(phoneNumber);
-      onCodeSent();
+      phoneAuthManager.update(() => phoneAuthManager.triggerOnCodeSent = true);
       return;
     }
     final completer = Completer<bool>();
@@ -142,6 +194,10 @@ class FirebaseAuthManager extends AuthManager
       timeout: Duration(seconds: 5),
       verificationCompleted: (phoneAuthCredential) async {
         await FirebaseAuth.instance.signInWithCredential(phoneAuthCredential);
+        phoneAuthManager.update(() {
+          phoneAuthManager.triggerOnCodeSent = false;
+          phoneAuthManager.phoneAuthError = null;
+        });
         // If you've implemented auto-verification, navigate to home page or
         // onboarding page here manually. Uncomment the lines below and replace
         // DestinationPage() with the desired widget.
@@ -151,15 +207,19 @@ class FirebaseAuthManager extends AuthManager
         // );
       },
       verificationFailed: (e) {
+        phoneAuthManager.update(() {
+          phoneAuthManager.triggerOnCodeSent = false;
+          phoneAuthManager.phoneAuthError = e;
+        });
         completer.complete(false);
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          content: Text('Error: ${e.message!}'),
-        ));
       },
       codeSent: (verificationId, _) {
-        _phoneAuthVerificationCode = verificationId;
+        phoneAuthManager.update(() {
+          phoneAuthManager.phoneAuthVerificationCode = verificationId;
+          phoneAuthManager.triggerOnCodeSent = true;
+          phoneAuthManager.phoneAuthError = null;
+        });
         completer.complete(true);
-        onCodeSent();
       },
       codeAutoRetrievalTimeout: (_) {},
     );
@@ -175,12 +235,14 @@ class FirebaseAuthManager extends AuthManager
     if (kIsWeb) {
       return _signInOrCreateAccount(
         context,
-        () => _webPhoneAuthConfirmationResult!.confirm(smsCode),
+        () => phoneAuthManager.webPhoneAuthConfirmationResult!.confirm(smsCode),
         'PHONE',
       );
     } else {
       final authCredential = PhoneAuthProvider.credential(
-          verificationId: _phoneAuthVerificationCode!, smsCode: smsCode);
+        verificationId: phoneAuthManager.phoneAuthVerificationCode!,
+        smsCode: smsCode,
+      );
       return _signInOrCreateAccount(
         context,
         () => FirebaseAuth.instance.signInWithCredential(authCredential),
