@@ -30,11 +30,16 @@ class FacultySportsLiveFirebaseUser extends BaseAuthUser {
     // Reloads the user when checking in order to get the most up to date
     // email verified status.
     if (loggedIn && !user!.emailVerified) {
-      FirebaseAuth.instance.currentUser
-          ?.reload()
-          .then((_) => user = FirebaseAuth.instance.currentUser);
+      refreshUser();
     }
     return user?.emailVerified ?? false;
+  }
+
+  @override
+  Future refreshUser() async {
+    await FirebaseAuth.instance.currentUser
+        ?.reload()
+        .then((_) => user = FirebaseAuth.instance.currentUser);
   }
 
   static BaseAuthUser fromUserCredential(UserCredential userCredential) =>
