@@ -86,7 +86,7 @@ String? serializeParam(
       case ParamType.DocumentReference:
         return _serializeDocumentReference(param as DocumentReference);
       case ParamType.Document:
-        final reference = (param as dynamic).reference as DocumentReference;
+        final reference = (param as FirestoreRecord).reference;
         return _serializeDocumentReference(reference);
 
       default:
@@ -182,9 +182,9 @@ enum ParamType {
 dynamic deserializeParam<T>(
   String? param,
   ParamType paramType,
-  bool isList, [
+  bool isList, {
   List<String>? collectionNamePath,
-]) {
+}) {
   try {
     if (param == null) {
       return null;
@@ -197,8 +197,8 @@ dynamic deserializeParam<T>(
       return paramValues
           .where((p) => p is String)
           .map((p) => p as String)
-          .map((p) =>
-              deserializeParam<T>(p, paramType, false, collectionNamePath))
+          .map((p) => deserializeParam<T>(p, paramType, false,
+              collectionNamePath: collectionNamePath))
           .where((p) => p != null)
           .map((p) => p! as T)
           .toList();
@@ -243,16 +243,16 @@ dynamic deserializeParam<T>(
 
 Future<dynamic> Function(String) getDoc(
   List<String> collectionNamePath,
-  Serializer serializer,
+  RecordBuilder recordBuilder,
 ) {
   return (String ids) => _deserializeDocumentReference(ids, collectionNamePath)
       .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .then((s) => recordBuilder(s));
 }
 
 Future<List<T>> Function(String) getDocList<T>(
   List<String> collectionNamePath,
-  Serializer<T> serializer,
+  RecordBuilder<T> recordBuilder,
 ) {
   return (String idsList) {
     List<String> docIds = [];
@@ -264,9 +264,7 @@ Future<List<T>> Function(String) getDocList<T>(
       docIds.map(
         (ids) => _deserializeDocumentReference(ids, collectionNamePath)
             .get()
-            .then(
-              (s) => serializers.deserializeWith(serializer, serializedData(s)),
-            ),
+            .then((s) => recordBuilder(s)),
       ),
     ).then((docs) => docs.where((d) => d != null).map((d) => d!).toList());
   };
