@@ -1,7 +1,10 @@
 // Export pages
 export '/pages/onboarding/onboarding_widget.dart' show OnboardingWidget;
-export '/pages/home/home_widget.dart' show HomeWidget;
+export '/pages/news/news_widget.dart' show NewsWidget;
 export '/pages/profile/profile_widget.dart' show ProfileWidget;
-export '/pages/edit_profile/edit_profile_widget.dart' show EditProfileWidget;
 export '/pages/forgotpassword/forgotpassword_widget.dart'
     show ForgotpasswordWidget;
+export '/pages/article_view/article_view_widget.dart' show ArticleViewWidget;
+export '/pages/app_settings/app_settings_widget.dart' show AppSettingsWidget;
+export '/pages/matches/matches_widget.dart' show MatchesWidget;
+export '/pages/teams/teams_widget.dart' show TeamsWidget;
