@@ -1,25 +1,24 @@
-import '/auth/firebase_auth/auth_util.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-class ForgotpasswordModel extends FlutterFlowModel {
+class AppSettingsModel extends FlutterFlowModel {
   ///  State fields for stateful widgets in this page.
 
-  // State field(s) for emailAddress widget.
-  TextEditingController? emailAddressController;
-  String? Function(BuildContext, String?)? emailAddressControllerValidator;
+  // State field(s) for SwitchListTile widget.
+  bool? switchListTileValue1;
+  // State field(s) for SwitchListTile widget.
+  bool? switchListTileValue2;
+  // State field(s) for SwitchListTile widget.
+  bool? switchListTileValue3;
 
   /// Initialization and disposal methods.
 
   void initState(BuildContext context) {}
 
-  void dispose() {
-    emailAddressController?.dispose();
-  }
+  void dispose() {}
 
   /// Action blocks are added here.
 
