@@ -1,24 +1,26 @@
-import '/auth/firebase_auth/auth_util.dart';
+import '/components/matcheslist/matcheslist_widget.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-class ForgotpasswordModel extends FlutterFlowModel {
+class MatchesModel extends FlutterFlowModel {
   ///  State fields for stateful widgets in this page.
 
-  // State field(s) for emailAddress widget.
-  TextEditingController? emailAddressController;
-  String? Function(BuildContext, String?)? emailAddressControllerValidator;
+  final unfocusNode = FocusNode();
+  // Model for matcheslist component.
+  late MatcheslistModel matcheslistModel;
 
   /// Initialization and disposal methods.
 
-  void initState(BuildContext context) {}
+  void initState(BuildContext context) {
+    matcheslistModel = createModel(context, () => MatcheslistModel());
+  }
 
   void dispose() {
-    emailAddressController?.dispose();
+    unfocusNode.dispose();
+    matcheslistModel.dispose();
   }
 
   /// Action blocks are added here.
